+++ lib/rubocop/cop/metrics/class_length.rb
@@ -29,6 +29,8 @@ module RuboCop
       #     HEREDOC
       #   end                 # 5 points
       #
+      #
+      # NOTE: This cop also applies for `Struct` definitions.
       class ClassLength < Base
         include CodeLength
 
@@ -37,16 +39,11 @@ module RuboCop
         end
 
         def on_casgn(node)
-          class_definition?(node) do
-            check_code_length(node)
-          end
+          _scope, _name, block_node = *node
+          check_code_length(node) if block_node.class_definition?
         end
 
         private
-
-        def_node_matcher :class_definition?, <<~PATTERN
-          (casgn nil? _ (block (send (const {nil? cbase} :Class) :new) ...))
-        PATTERN
 
         def message(length, max_length)
           format('Class has too many lines. [%<length>d/%<max>d]',